requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
alpaca-py>=0.21.0
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0
pyarrow
